+++ packages/gatsby-theme-blog-core/gatsby-config.js
@@ -9,7 +9,7 @@ module.exports = themeOptions => {
       title: `Blog Title Placeholder`,
       author: `Name Placeholder`,
       description: `Description placeholder`,
-      siteURL: 'https://example.com',
+      siteUrl: 'https://example.com',
       social: [
         {
           name: `Twitter`,