+++ packages/gatsby/src/redux/reducers/index.js
@@ -14,6 +14,11 @@ import { lastAction } from "./last-action"
 import { jobsV2Reducer } from "./jobsv2"
 import { flattenedPluginsReducer } from "./flattened-plugins"
 import { resolvedNodesCacheReducer } from "./resolved-nodes"
+import { pageDataStatsReducer } from "./page-data-stats"
+import { componentsReducer } from "./components"
+import { componentDataDependenciesReducer } from "./component-data-dependencies"
+import { babelrcReducer } from "./babelrc"
+import { jobsReducer } from "./jobs"
 
 /**
  * @property exports.nodesTouched Set<string>
@@ -30,19 +35,19 @@ module.exports = {
   schema: schemaReducer,
   pages: pagesReducer,
   status: statusReducer,
-  componentDataDependencies: require(`./component-data-dependencies`),
-  components: require(`./components`),
+  componentDataDependencies: componentDataDependenciesReducer,
+  components: componentsReducer,
   staticQueryComponents: staticQueryComponentsReducer,
-  jobs: require(`./jobs`),
+  jobs: jobsReducer,
   jobsV2: jobsV2Reducer,
   webpack: webpackReducer,
   webpackCompilationHash: webpackCompilationHashReducer,
   redirects: redirectsReducer,
-  babelrc: require(`./babelrc`),
+  babelrc: babelrcReducer,
   schemaCustomization: require(`./schema-customization`),
   themes: themesReducer,
   logs: logReducer,
   inferenceMetadata: require(`./inference-metadata`),
-  pageDataStats: require(`./page-data-stats`),
+  pageDataStats: pageDataStatsReducer,
   pageData: pageDataReducer,
 }