+++ packages/gatsby-plugin-typescript/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.4.3](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-typescript@2.4.2...gatsby-plugin-typescript@2.4.3) (2020-05-20)
+
+**Note:** Version bump only for package gatsby-plugin-typescript
+
 ## [2.4.2](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-typescript@2.4.1...gatsby-plugin-typescript@2.4.2) (2020-05-05)
 
 **Note:** Version bump only for package gatsby-plugin-typescript